requirements.txt
Django>=5.0,<6
django-htmx>=1.21
Jinja2>=3.1
django-environ>=0.11
psycopg2-binary>=2.9
celery[redis]>=5.3
redis>=5.0
whitenoise>=6.6
cryptography>=42.0
requests>=2.31
dnspython>=2.6
djangorestframework>=3.15,<4
gunicorn>=22.0
